Fix camera extra settings menu
If the list of settings entries is scrolled, the adapter returns a
reused View instead of creating a new one. This makes the last
entries unusable. This has been broken for ages but only shows up
if the camera has more than a few settings (which the stock camera
doesn't)

http://gerrit.sudoservers.com/#/c/7255/1

Change-Id: Ib78f690fb1595792b11bba622b718f05d9251761